applicationControllers-added

+++ controllers/AuthController.js
@@ -4,7 +4,9 @@ const Student = require("../models/StudentModel");
 const OtpModel = require("../models/OtpModel");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-const emailGenerator = require("../utils/EmailGenerator");
+const nodemailer = require("nodemailer");
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
 const authCtrl = {};
 
 const expiryAccessToken = "1h";
@@ -43,8 +45,6 @@ const generateOTP = ()=>{
 authCtrl.Login = async (req, res) => {
     const email  = req.body.email;
     console.log("email",email)
-
-    if(!email || !req.body.password) return res.status(400).json({msg:"Invalid Email or Password"})
 
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     if (!emailRegex.test(email)) return res.status(400).json({ msg: "Invalid Email format" });
@@ -96,7 +96,7 @@ authCtrl.Login = async (req, res) => {
 authCtrl.regenerateAccessToken = async (req, res) => {
     const refreshToken = req.cookies.refresh_token;
 
-    if (refreshToken == null) return res.sendStatus(400);
+    if (typeof refreshToken !== 'string') return res.sendStatus(400);
 
     jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, (err, user) => {
         if (err) return res.sendStatus(400)
@@ -119,9 +119,13 @@ authCtrl.Logout = async (req, res) => {
     res.sendStatus(204)
 }
 
+//Sent OTPs to The Mail id
 authCtrl.SendOTP = async(req,res)=>{
     const email = req.body.email;
     console.log(email)
+
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!emailRegex.test(email)) return res.status(400).json({ msg: "Invalid Email format" });
 
     const student = await Student.findOne({email}).lean();
     const employee = await Employee.findOne({email}).lean();
@@ -141,13 +145,32 @@ authCtrl.SendOTP = async(req,res)=>{
         const otpDoc = new OtpModel({email,otp:OTP})
         await otpDoc.save();
 
-        const link = "https://realify.online";
-
-        const result = emailGenerator(link, email);
-
-        if(!result) return res.status(500).json({msg:"Couldn't sent OTP"})
-
-        res.status(200).json({msg:'Otp Sent successfully'})
+        const transporter = nodemailer.createTransport({
+            host:"smtp.gmail.com",
+            port:465,
+            secure:true,
+            auth: {
+              user:process.env.MAIL_USER,
+              pass:process.env.MAIL_PASSWORD
+            }
+          });
+        
+          const mailOptions = {
+            from:process.env.MAIL_USER,
+            to: email,
+            subject: "OTP to Change Password",
+            text: OTP
+          };
+        
+          transporter.sendMail(mailOptions, (error, info)=> {
+            if (error) {
+              console.log(error);
+              return res.status(500).json({msg:"Couldn't Sent OTP "});
+            } else {
+              console.log('Email sent: ' + info.response);
+              return res.status(200).json({msg:"OTP Sent Successfully"})
+            }
+          });
 
     }catch(error){
         console.error(error);
@@ -155,17 +178,32 @@ authCtrl.SendOTP = async(req,res)=>{
     }
 }
 
-authCtrl.VerifyOTP = async(req,res)=>{
+
+//Verify Email id;
+authCtrl.VerifyMail = async(req,res)=>{
     const otp = req.body.otp;
     const email = req.body.email;
-    const validOtp = await OtpModel.findOne({email,otp})
-    if(!validOtp){
-        return res.status(403).json({msg:'Invalid Otp number'})
-    }
 
-    await OtpModel.findByIdAndDelete(validOtp._id)
+    if(typeof otp !== 'string') return res.status(400).json({msg:"Invalid OTP format"})
     
-    return res.status(200).json({msg:"Email verified"});
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!emailRegex.test(email)) return res.status(400).json({ msg: "Invalid Email format" });
+    
+    try {
+        const validOtp = await OtpModel.findOne({email,otp});
+        
+        if(!validOtp){
+            return res.status(403).json({msg:'Invalid Otp number'})
+        }
+    
+        await OtpModel.findByIdAndDelete(validOtp._id)
+        
+        res.status(200).json({msg:"Email verified"});
+        
+    } catch (error) {
+        res.status(500).json({msg:"Something went wrong"});
+        
+    }
 }
 
 module.exports = authCtrl;
